dodate verzije za red voznje i linije

+++ jgspApp/api/controllers/stationCont.js
@@ -5,39 +5,49 @@ var Line = mongoose.model('line');
 module.exports.removeStation = function(req, res)
 {
     
-    if(!req.params._id ) {
-        sendJSONresponse(res, 400, {
-            "message": "All fields required"
-        });
-        return;
+    if(!req.params._id) {
+
+        return res.status(400).json({"message" : "You have to click on station you want to remove!"})
+       
     }
+    Station.findById(req.params._id).then(abc => {
+    
+        if(abc == null || abc == undefined)
+        {
+            return res.status(404).json({"message" : "Station that you are trying to remove either do not exist or was previously deleted by another user."});
+        }else {
+            Station.findOneAndRemove({_id: req.params._id}).then(bla =>{
+                // if(bla == null || bla == udefined)
+                // {
+                //     return res.status(404).json({"message": "Station that you are trying to delete either do not exist or was previously deleted by another user."})
+                // }else {
+                    Line.find().then(aa => {
+                        aa.forEach(bb => {
+                            bb.stations.forEach(cc => {
+                                if(cc == req.params._id)
+                                {
+                                    bb.stations.remove(cc);
+                                    Line.findOneAndUpdate({_id: bb._id}, {stations:bb.stations}).then(abc => {});
+                                }
+                            })
+                        })
+                    })
+                   return res.status(200).json({
+                        "message" : "Station successfully removed."
+                    });
+               // }
+                
+            });
+        }
+    })
 
-    Station.findOneAndRemove({_id: req.params._id}).then(bla =>{
-
-        Line.find().then(aa => {
-            aa.forEach(bb => {
-                bb.stations.forEach(cc => {
-                    if(cc == req.params._id)
-                    {
-                        bb.stations.remove(cc);
-                        Line.findOneAndUpdate({_id: bb._id}, {stations:bb.stations}).then(abc => {});
-                    }
-                })
-            })
-        })
-        res.status(200).json({
-            "message" : "Station successfully removed."
-    });
-});
+    
 }
 
 
 module.exports.changeStation = function(req, res){
     if(!req.body.Name || !req.body.Address || !req.body.Latitude || !req.body.Longitude ) {
-        sendJSONresponse(res, 400, {
-            "message": "All fields required"
-        });
-        return;
+        return res.status(400).json({"message" : "You have to move station you want to change!"})
     }
 
     Station.findById(req.body.Id).exec().then(st => {
